Wagtail + Google images service
Sample code for a Wagtail custom image model that use the Google images
service to serve images (instead of generating renditions and serving
them from storage).

Most of the image transformations are implemented with the equivalent
Google URL parameters, but not all.

This code is a bit weird in that it never actually saves the rendition
instances in the database, since the Google images service handles all
the transformations for you, and is incredibly fast. Not saving things
in the database probably breaks how some Wagtail sites use images.